doc: add roadmap for features
Currently any client with correct server path would be accepted by the
server. It's apparently not safe and is prone for attacks. ZAP is the
ZeroMQ Authentication Protocol, it allows for an artitecture where
server would only accept known clients with their public certificates.

A ping for every editor input is not enough, more datas about my own
coding process should be collected and processed.