fix: set the correct timestamp for an expired session
An expired session's end property should not be the time it's closed.
User can stop editing the file and never leave the buffer for a very
long time. Then all those idle time after the last tick should not be
counted for the total editing time. Instead, a better way to mark the
end property of the session should be one minute after the last tick.